% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fregion.rect.R
\name{plot.fregion.rect}
\alias{plot.fregion.rect}
\title{Visualizes rectangular confidence regions achieved from \link{fregion.rect}.}
\usage{
\method{plot}{fregion.rect}(rect, npc = NULL, blegend = TRUE,
  returntype = NULL, grid.size = 200, ...)
}
\arguments{
\item{rect}{A 'fregion.rect' object, the output from \link{fregion.rect} funciton.}

\item{npc}{Number of fPCs to plot.}

\item{grid.size}{Since this function makes plots, fd objects are evaluated and gird.size determines how find the grid should be.}
}
\description{
Visualizes rectangular confidence regions achieved from \link{fregion.rect}.
}
